Better testing for sites.
This pull-request moves the code to find the most recent-entries
from the "driver" into the "site".  As part of that I've overhauled
the API a little, such that creating a new site-object parses both
the comments and the blog-posts themselves.

This cuts down on repetition on the caller, and allows me to add
some simple tests.

Changing the API isn't something I'd do lightly, but this is an
internal package with no external users so I think it is safe enough,
and the added test-coverage means the change is worthwhile.

This closes #24.